requirements.txt
addict==2.4.0
chardet==4.0.0
cycler==0.10.0
dataclasses==0.8
importlib-metadata==4.8.1
kiwisolver==1.3.1
matplotlib==3.3.4
mmcv-full==1.3.9
mmsegmentation==0.17.0
numpy==1.19.5
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
prettytable==2.2.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.62.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
yapf==0.31.0
zipp==3.5.1